Fix TS/Python errors, rename activity files, remove ChannelData
- Fix double-rename bug: CoreCoreActivity -> CoreActivity in conversation clients
- Rename activity.ts -> core-activity.ts, activity.py -> core_activity.py,
  test_activity.py -> test_core_activity.py, activity.spec.ts -> core-activity.spec.ts
- Remove ChannelData files (node, python, dotnet) — no longer needed
- Flatten CoreActivity<TChannelData> generic to CoreActivity in dotnet
- Fix ruff import sort lint errors in Python
- All tests pass: 25 Python, 16 dotnet, node build clean
