--- dotnet/src/Botas/Schema/ChannelData.cs
@@ -1,12 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace Botas.Schema;
-
-public class ChannelData()
-{
-    [JsonPropertyName("clientActivityID")]
-    public string? ClientActivityId { get; set; }
-
-    [JsonExtensionData]
-    public ExtendedPropertiesDictionary Properties { get; set; } = new();
-}
