+++ node/packages/botas/src/schema/core-activity.ts
