+++ node/packages/botas/src/schema/index.ts
@@ -1,6 +1,5 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-export * from './activity.js'
+export * from './core-activity.js'
 export * from './activity-type.js'
-export * from './channel-data.js'
