--- python/packages/botas/src/botas/schema/channel_data.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-from botas.schema.activity import _CamelModel
-
-
-class ChannelData(_CamelModel):
-    client_activity_id: str | None = None
