+++ python/packages/botas/src/botas/schema/core_activity.py
